receivers take proper args and check for 'raw' kwarg

+++ imagersite/imager_profile/handlers.py
@@ -1,21 +1,25 @@
-from django.contrib.auth import User
+from django.contrib.auth.models import User
 from django.dispatch import receiver
 from django.db.models.signals import post_save, post_delete
 from models import ImagerProfile
 
 
 @receiver(post_save, sender=User)
-def auto_create_profile(sender, instance, **kwargs):
-    try:
-        instance.profile
-    except ImagerProfile.DoesNotExist:
-        instance.profile = ImagerProfile()
-        instance.profile.save()
+def auto_create_profile(sender, **kwargs):
+    if not kwargs['raw']:
+        try:
+            instance = kwargs['instance']
+            instance.profile
+        except ImagerProfile.DoesNotExist:
+            instance.profile = ImagerProfile()
+            instance.profile.save()
 
 
 @receiver(post_delete, sender=User)
-def auto_delete_profile(sender, instance, **kwargs):
-    try:
-        instance.profile.delete()
-    except ImagerProfile.DoesNotExist:
-        pass
+def auto_delete_profile(sender, **kwargs):
+    if not kwargs['raw']:
+        try:
+            instance = kwargs['instance']
+            instance.profile.delete()
+        except ImagerProfile.DoesNotExist:
+            pass
